chess-backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
python-chess
pydantic
starlette
requests
aiofiles
